+++ tech_reader/techdesignrule.h
@@ -13,16 +13,16 @@ public:
 	TechDesignRule();
 	void setName(QString);
 	QString getName();
-	void setMinimumWidth(int i);
+	void setMinimumWidth(double i);
 	void setWidthMessage(QString m);
 	QString getWidthMessage();
 	void setSpacing(QString name, int d, QString message, bool touching_ok);
 	QStringList getSpacingRules();
 	TechDesignRuleSpacing getSpacingRule(QString k);
-	int getMinimumWidth();
+	qreal getMinimumWidth();
 
 private:
-	int m_minWidth;
+	qreal m_minWidth;
 	QString m_name;
 	QString m_widthMessage;
 	QMap<QString,TechDesignRuleSpacing> m_spacingList;
